Bring Up to Date and Fix Indicator Glitch
Took a long break with this project and apparently didn't push the changes I made before taking the break because it has glitches. There are still glitches in this push, but it's more up-to-date and generally has a better code structure imo. Next glitch to fix is related to being in check.